requirements.txt
seleniumbase
pyyaml
requests
beautifulsoup4